Split Apollo page into server and client side components (server side is currently just the tabs, but making this change for future maintainability and scalability)
